removed duplicate SingletonMeta definition
